pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agente-email"
version = "0.1.0"
description = "Dockerized Thunderbird-Mail AI Agent (IMAP + Google Calendar + LangGraph + OpenRouter)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12",
  "google-api-python-client>=2.130",
  "google-auth>=2.29",
  "google-auth-oauthlib>=1.2",
  "langchain>=0.2",
  "langgraph>=0.2",
  "openai>=1.40",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "python-dateutil>=2.9",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.10",
  "pytest>=8.2",
  "pytest-mock>=3.14",
  "ruff>=0.6",
  "types-python-dateutil>=2.9.0.20240316",
]

[project.scripts]
agente-email = "agent.main:main"
agente-email-auth-google = "agent.auth_google:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "preserve"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
no_implicit_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
mypy_path = "src"
packages = ["agent"]

[[tool.mypy.overrides]]
module = [
  "googleapiclient.*",
  "google_auth_oauthlib.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
